Consolidate FlyDSL forward into flydsl.py alongside native BwOp
Move flydslF forward op from flydsl_forward.py into flydsl.py (which upstream
__init__/dispatch/tests/bench references. Forward keeps op_bw=None -> dispatch
(ck.BwOp); native flydsl fw+bw pairing is not yet f16-verified.
