+++ mslk/attention/fmha/dispatch.py
@@ -14,16 +14,7 @@ from typing import Any, List, Optional, Sequence, Tuple, Type, TypeVar
 
 import torch
 
-from . import (
-    attn_bias,
-    ck,
-    cutlass,
-    flash,
-    flash3,
-    flash_mtia,
-    flydsl_forward,
-    triton_splitk,
-)
+from . import attn_bias, ck, cutlass, flash, flash3, flash_mtia, flydsl, triton_splitk
 from .common import AttentionBwOpBase, AttentionFwOpBase, Inputs
 
 T = TypeVar("T", Type[AttentionFwOpBase], Type[AttentionBwOpBase])
@@ -143,7 +134,7 @@ def _dispatch_fw_priority_list(
                 ck.FwOp,
                 # After CK so it does not displace the default, but available for
                 # inputs CK declines or when explicitly selected.
-                flydsl_forward.FwOp,
+                flydsl.FwOp,
             ]
         )
     # pyrefly: ignore [bad-argument-type]
